fixed errors with walls and connections
walls do not need to remove 1 from width or height when converting.

intersections so not need to remove 1 from width or height when converting.

When converting vertical to horizontal walls/blocking edges, the points must be switched, so that the points go left to right.

when converting horizontal to vertical walls/edges, the points must be switched, so that the points go from top to bottom.